Extract views
Extract ViewXx classes from ViewXx classess.
